feat: delete rating operation
