+++ recipelib/operations/recipes/recipe_actions.py
@@ -11,8 +11,7 @@ from recipelib.models import (
     RecipeTag,
     Tag,
 )
-from recipelib.models.rating import Rating
-from recipelib.serializers import RatingSerializer, RecipeSerializer
+from recipelib.serializers import RecipeSerializer
 from recipelib.utils import error_json_response, not_found_json_response
 
 schema = {
@@ -260,28 +259,3 @@ def get_feed(req):
     except Exception as err:
         print(err)
         return error_json_response(err)
-
-
-def rate_recipe(req, data, recipe):
-    try:
-        rating = Rating.objects.get_or_create(recipe=recipe, user=req.user)[0]
-        print("domo")
-        rating.like = data["like"]
-        rating.save()
-        return JsonResponse(
-            RatingSerializer(rating).data,
-            safe=False,
-            status=201,
-        )
-    except Exception as err:
-        print(err)
-        return error_json_response(err)
-
-
-rate_recipe.schema = {
-    "type": "object",
-    "properties": {
-        "like": {"type": "boolean"},
-    },
-    "required": ["like"],
-}
